Annotate fix with pcr
This fixes an issue with some annotatation/features being ‘cut’ from
the original template during a PCR reaction. This was due to the
template being spliced from the 3’ ends of the primers and appended to
the primers, which removes any features associated at the 3’ and 5’ end
of the fragment. Template will not be spliced at the 5’ of the
annealing portion of the primer during the PCR reaction.